+++ js/image/Index.svelte
@@ -42,7 +42,7 @@
 	let fullscreen = $state(false);
 	let dragging = $state(false);
 	let active_source = $derived.by(() =>
-		gradio.props.sources ? gradio.props.sources[0] : null,
+		gradio.props.sources ? gradio.props.sources[0] : null
 	);
 
 	let upload_component: ImageUploader;
