+++ js/imageeditor/InteractiveImageEditor.svelte
@@ -28,7 +28,7 @@
 		type LayerOptions,
 		type Transform,
 		type Source,
-		type WebcamOptions,
+		type WebcamOptions
 	} from "./shared/types";
 
 	export let brush: IBrush;
@@ -91,27 +91,27 @@
 		const bg = blobs.background
 			? upload(
 					await prepare_files([new File([blobs.background], "background.png")]),
-					root,
+					root
 				)
 			: Promise.resolve(null);
 
 		const layers = blobs.layers
 			.filter(is_not_null)
 			.map(async (blob, i) =>
-				upload(await prepare_files([new File([blob], `layer_${i}.png`)]), root),
+				upload(await prepare_files([new File([blob], `layer_${i}.png`)]), root)
 			);
 
 		const composite = blobs.composite
 			? upload(
 					await prepare_files([new File([blobs.composite], "composite.png")]),
-					root,
+					root
 				)
 			: Promise.resolve(null);
 
 		const [background, composite_, ...layers_] = await Promise.all([
 			bg,
 			composite,
-			...layers,
+			...layers
 		]);
 
 		return {
@@ -119,7 +119,7 @@
 			layers: layers_
 				.flatMap((layer) => (Array.isArray(layer) ? layer : [layer]))
 				.filter(is_file_data),
-			composite: Array.isArray(composite_) ? composite_[0] : composite_,
+			composite: Array.isArray(composite_) ? composite_[0] : composite_
 		};
 	}
 
@@ -162,14 +162,14 @@
 				id,
 				"background",
 				new File([blobs.background], "background.png"),
-				null,
+				null
 			]);
 		if (blobs.composite)
 			images.push([
 				id,
 				"composite",
 				new File([blobs.composite], "composite.png"),
-				null,
+				null
 			]);
 		blobs.layers.forEach((layer, i) => {
 			if (layer)
@@ -177,16 +177,16 @@
 					id as string,
 					`layer`,
 					new File([layer], `layer_${i}.png`),
-					i,
+					i
 				]);
 		});
 		await Promise.all(
 			images.map(async ([image_id, type, data, index]) => {
 				return accept_blobs({
 					binary: true,
-					data: { file: data, id: image_id, type, index },
+					data: { file: data, id: image_id, type, index }
 				});
-			}),
+			})
 		);
 		image_id = id;
 		dispatch("change");
