+++ js/imageeditor/shared/ImageEditor.svelte
@@ -154,7 +154,7 @@
 						_type: string;
 					};
 			  }
-			| any,
+			| any
 	): Promise<void> {
 		if (!editor || !source || !check_if_should_init()) return;
 		let url: string;
@@ -189,7 +189,7 @@
 	 * @returns {Promise<string | null>} - The ID of the created layer, or null if failed
 	 */
 	export async function add_layers_from_url(
-		source: FileData[] | any,
+		source: FileData[] | any
 	): Promise<void> {
 		if (!editor || !source.length || !check_if_should_init()) return;
 
@@ -237,7 +237,7 @@
 
 				last_dimensions = {
 					width: current_dimensions.width,
-					height: current_dimensions.height,
+					height: current_dimensions.height
 				};
 			}
 		}
@@ -329,7 +329,7 @@
 			fixed_canvas,
 			border_region,
 			layer_options,
-			theme_mode,
+			theme_mode
 		});
 
 		brush.on("change", () => {
@@ -346,7 +346,7 @@
 			dark: true,
 			fixed_canvas: false,
 			border_region: 0,
-			pad_bottom: 40,
+			pad_bottom: 40
 		});
 
 		editor.scale.subscribe((_scale) => {
@@ -422,7 +422,7 @@
 	 * @param {File[]} files - The uploaded files
 	 */
 	async function handle_files(
-		files: File[] | Blob[] | File | Blob | null,
+		files: File[] | Blob[] | File | Blob | null
 	): Promise<void> {
 		if (files == null) return;
 		if (!sources.includes("upload")) return;
@@ -460,7 +460,7 @@
 	 */
 	function handle_subtool_change({
 		tool,
-		subtool,
+		subtool
 	}: {
 		tool: ToolbarTool;
 		subtool: Subtool | null;
@@ -557,7 +557,7 @@
 					brush_options.colors.find((color) =>
 						Array.isArray(color)
 							? color[0] === brush_options.default_color
-							: color === brush_options.default_color,
+							: color === brush_options.default_color
 					) || brush_options.colors[0];
 
 				color_value = Array.isArray(default_color)
@@ -567,15 +567,15 @@
 				color_value = selected_color;
 			}
 			return color_value;
-		})(),
+		})()
 	);
 
 	$: brush?.set_brush_size(
-		typeof selected_size === "number" ? selected_size : 25,
+		typeof selected_size === "number" ? selected_size : 25
 	);
 
 	$: brush?.set_eraser_size(
-		typeof selected_eraser_size === "number" ? selected_eraser_size : 25,
+		typeof selected_eraser_size === "number" ? selected_eraser_size : 25
 	);
 
 	$: disable_click =
@@ -597,7 +597,7 @@
 		zoom.set_zoom(
 			direction === "in"
 				? zoom_level + (zoom_level < 1 ? 0.1 : zoom_level * 0.1)
-				: zoom_level - (zoom_level < 1 ? 0.1 : zoom_level * 0.1),
+				: zoom_level - (zoom_level < 1 ? 0.1 : zoom_level * 0.1)
 		);
 	}
 
@@ -634,7 +634,7 @@
 
 		await editor.add_image({
 			image,
-			resize: false,
+			resize: false
 		});
 		handle_subtool_change({ tool: "image", subtool: null });
 		dispatch("change");
@@ -683,7 +683,7 @@
 		on_drag_change: (dragging) => (is_dragging = dragging),
 		on_files: handle_files,
 		accepted_types: "image/*",
-		disable_click: disable_click,
+		disable_click: disable_click
 	}}
 	aria-label={"Click to upload or drop files"}
 	aria-dropeffect="copy"
