+++ js/utils/src/utils.svelte.ts
@@ -2,8 +2,42 @@ import type { ActionReturn } from "svelte/action";
 import type { Client } from "@gradio/client";
 import type { ComponentType, SvelteComponent } from "svelte";
 import { getContext, tick, untrack } from "svelte";
-import { type SharedProps } from "@gradio/core";
 import type { Component } from "svelte";
+
+export interface SharedProps {
+	elem_id?: string;
+	elem_classes: string[];
+	components?: string[];
+	server_fns?: string[];
+	interactive: boolean;
+	visible: boolean | "hidden";
+	id: number;
+	container: boolean;
+	target: HTMLElement;
+	theme_mode: "light" | "dark" | "system";
+	version: string;
+	root: string;
+	autoscroll: boolean;
+	max_file_size: number | null;
+	formatter: any; //I18nFormatter;
+	client: Client;
+	scale: number;
+	min_width: number;
+	padding: number;
+	load_component: (
+		arg0: string,
+		arg1: "base" | "example" | "component"
+	) => LoadingComponent; //component_loader;
+	loading_status?: any;
+	label: string;
+	show_label: boolean;
+	validation_error?: string | null;
+	theme?: "light" | "dark";
+	show_progress: boolean;
+	api_prefix: string;
+	server: ServerFunctions;
+	attached_events?: string[];
+}
 
 export type LoadingComponent = Promise<{
 	default: Component;
